+++ src/Shape.ts
@@ -10,6 +10,14 @@ var HAS_SHADOW = 'hasShadow';
 var SHADOW_RGBA = 'shadowRGBA';
 
 // TODO: cache gradient from context
+// TODO: write a test for adding destroyed shape into the layer
+// will it draw?
+// will it pass hit test?
+// show warning on adding destroyed shape?
+
+// TODO: idea - use only "remove" (or destroy method)
+// how? on add, check that every inner shape has reference in konva store with color
+// on remove - clear that reference
 
 function _fillFunc(context) {
   context.fill();